+++ spinn_front_end_common/utilities/helpful_functions.py
@@ -18,14 +18,11 @@ import logging
 import struct
 from spinn_utilities.log import FormatAdapter
 from spinn_machine import CoreSubsets
-from spinnman.model import ExecutableTargets
 from spinnman.model.enums import CPUState
-from spinnman.model.cpu_infos import CPUInfos
 from data_specification import utility_calls
 from spinn_front_end_common.abstract_models import AbstractHasAssociatedBinary
 from spinn_front_end_common.utilities.exceptions import ConfigurationException
 from spinn_front_end_common.utilities.utility_objs import ExecutableType
-from .globals_variables import get_simulator
 from .constants import BYTES_PER_WORD, MICRO_TO_MILLISECOND_CONVERSION
 
 logger = FormatAdapter(logging.getLogger(__name__))
@@ -327,92 +324,3 @@ def find_executable_start_type(machine_vertex):
         return machine_vertex.get_binary_start_type()
     else:
         return None
-
-
-def _emergency_state_check(txrx, app_id):
-    """
-    :param ~.Transceiver txrx:
-    :param int app_id:
-    """
-    # pylint: disable=broad-except
-    try:
-        rte_count = txrx.get_core_state_count(
-            app_id, CPUState.RUN_TIME_EXCEPTION)
-        watchdog_count = txrx.get_core_state_count(app_id, CPUState.WATCHDOG)
-        if rte_count or watchdog_count:
-            logger.warning(
-                "unexpected core states (rte={}, wdog={})",
-                txrx.get_cores_in_state(None, CPUState.RUN_TIME_EXCEPTION),
-                txrx.get_cores_in_state(None, CPUState.WATCHDOG))
-    except Exception:
-        logger.exception(
-            "Could not read the status count - going to individual cores")
-        machine = txrx.get_machine_details()
-        infos = CPUInfos()
-        errors = list()
-        for chip in machine.chips:
-            for p in chip.processors:
-                try:
-                    info = txrx.get_cpu_information_from_core(
-                        chip.x, chip.y, p)
-                    if info.state in (
-                            CPUState.RUN_TIME_EXCEPTION, CPUState.WATCHDOG):
-                        infos.add_processor(chip.x, chip.y, p, info)
-                except Exception:
-                    errors.append((chip.x, chip.y, p))
-        logger.warning(txrx.get_core_status_string(infos))
-        logger.warning("Could not read information from cores {}".format(
-            errors))
-
-
-# TRICKY POINT: Have to delay the import to here because of import circularity
-def _emergency_iobuf_extract(txrx, executable_targets):
-    """
-    :param ~.Transceiver txrx:
-    :param ExecutableTargets executable_targets:
-    """
-    # pylint: disable=protected-access
-    from spinn_front_end_common.interface.interface_functions import (
-        ChipIOBufExtractor)
-    sim = get_simulator()
-    extractor = ChipIOBufExtractor(
-        recovery_mode=True, filename_template="emergency_iobuf_{}_{}_{}.txt")
-    extractor(txrx, executable_targets, sim._executable_finder,
-              sim._app_provenance_file_path, sim._system_provenance_file_path)
-
-
-def emergency_recover_state_from_failure(txrx, app_id, vertex, placement):
-    """ Used to get at least *some* information out of a core when something\
-        goes badly wrong. Not a replacement for what abstract spinnaker base\
-        does.
-
-    :param ~spinnman.transceiver.Transceiver txrx: The transceiver.
-    :param int app_id: The ID of the application.
-    :param AbstractHasAssociatedBinary vertex:
-        The vertex to retrieve the IOBUF from if it is suspected as being dead
-    :param ~pacman.model.placements.Placement placement:
-        Where the vertex is located.
-    """
-    # pylint: disable=protected-access
-    _emergency_state_check(txrx, app_id)
-    target = ExecutableTargets()
-    path = get_simulator()._executable_finder.get_executable_path(
-        vertex.get_binary_file_name())
-    target.add_processor(
-        path, placement.x, placement.y, placement.p,
-        vertex.get_binary_start_type())
-    _emergency_iobuf_extract(txrx, target)
-
-
-def emergency_recover_states_from_failure(txrx, app_id, executable_targets):
-    """ Used to get at least *some* information out of a core when something\
-        goes badly wrong. Not a replacement for what abstract spinnaker base\
-        does.
-
-    :param ~spinnman.transceiver.Transceiver txrx: The transceiver.
-    :param int app_id: The ID of the application.
-    :param ~spinnman.model.ExecutableTargets executable_targets:
-        The what/where mapping
-    """
-    _emergency_state_check(txrx, app_id)
-    _emergency_iobuf_extract(txrx, executable_targets)
