+++ spinn_front_end_common/utility_models/extra_monitor_support_machine_vertex.py
@@ -34,11 +34,13 @@ from spinn_front_end_common.utilities.utility_objs.\
 from spinn_front_end_common.utilities.constants import (
     SARK_PER_MALLOC_SDRAM_USAGE, DATA_SPECABLE_BASIC_SETUP_INFO_N_BYTES,
     BYTES_PER_WORD, BYTES_PER_KB)
+from spinn_front_end_common.utilities.helpful_functions import (
+    convert_vertices_to_core_subset)
+from spinn_front_end_common.utilities.emergency_recovery import (
+    emergency_recover_state_from_failure)
 from .data_speed_up_packet_gatherer_machine_vertex import (
     DataSpeedUpPacketGatherMachineVertex as
     Gatherer)
-from spinn_front_end_common.utilities.helpful_functions import (
-    convert_vertices_to_core_subset, emergency_recover_state_from_failure)
 
 log = FormatAdapter(logging.getLogger(__name__))
